feat: logic opt & update: README

+++ extensions/simpletexocr/src/index.tsx
@@ -23,9 +23,35 @@ export default async () => {
     "Content-Type": "multipart/form-data",
     token: token,
   };
-  const res = await axios.post("https://server.simpletex.cn/api/latex_ocr", formData, { headers });
-  await Clipboard.copy(res.data.res.latex);
-  // remove file
-  fs.unlinkSync(savePath);
-  await showHUD("✅ OCR 成功");
+  try {
+    const res = await axios.post("https://server.simpletex.cn/api/latex_ocr", formData, { headers });
+    console.log(res);
+    const data = res.data;
+    if (!data.status) {
+      throw new Error("API 报错");
+    }
+    await Clipboard.copy(data.res.latex);
+    await showHUD("✅ OCR 成功");
+    fs.unlinkSync(savePath);
+  } catch (err) {
+    if (axios.isAxiosError(err) && err.response) {
+      const err_dict: { [key: number]: string } = {
+        401: "鉴权失败",
+        402: "没有可以用以调用接口的资源，如无资源包或账户余额不足",
+        404: "找不到对应的API或对应的版本",
+        405: "错误的请求方法",
+        413: "图片文件过大",
+        429: "超出最大调用的并发请求量，请稍后再试",
+        500: "没有文件导致的服务器错误",
+        503: "服务器未启动/维护中",
+      };
+      const status_code: number = err.response.status;
+      const err_msg = err_dict[status_code] || "未知错误";
+      await showHUD("❌ 请求 API 失败：" + err_msg);
+    } else {
+      await showHUD("❌ 请求 API 失败");
+    }
+    fs.unlinkSync(savePath);
+    return;
+  }
 };
